requirements.txt
openai
pandas
openpyxl
numpy
python-dotenv
